backend/pyproject.toml
[project]
name = "myvpn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "gunicorn>=23.0.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-jose>=3.5.0",
    "python-json-logger>=4.0.0",
    "redis>=7.1.0",
    "uuid-utils>=0.12.0",
    "uvicorn>=0.40.0",
]
